+++ src/server/db/queries.ts
@@ -1,0 +1,33 @@
+import { db } from "~/server/db";
+import { eq } from "drizzle-orm";
+import {
+  files_table as filesSchema,
+  folders_table as foldersSchema,
+} from "~/server/db/schema";
+
+export async function getAllParentsForFolder(folderId: number) {
+  const parents = [];
+  let currentId: number | null = folderId;
+  while (currentId !== null) {
+    const folder = await db
+      .select()
+      .from(foldersSchema)
+      .where(eq(foldersSchema.id, currentId));
+    if (!folder[0]) {
+      throw new Error("Folder not found");
+    }
+    parents.unshift(folder[0]);
+    currentId = folder[0].parent;
+  }
+  return parents;
+}
+
+export function getFolders(folderId: number) {
+  return db
+    .select()
+    .from(foldersSchema)
+    .where(eq(foldersSchema.parent, folderId));
+}
+export function getFiles(folderId: number) {
+  return db.select().from(filesSchema).where(eq(filesSchema.parent, folderId));
+}
